+++ packages/core/src/docs/08-theming.stories.tsx
@@ -278,6 +278,7 @@ The global theme is provided by the DataEditor by default and can be overriden b
 | editorFontSize | --gdg-editor-font-size | string | The font size used by overlay editors. |
 | lineHeight | None | number | A unitless scaler which defines the height of a line of text relative to the ink size. |
 | checkboxMaxSize | --gdg-checkbox-max-size | number | The maximum size of checkboxes (in pixels), e.g. for boolean cell and row markers. |
+| checkboxIndeterminateInnerSize | None | number | The size of the inner square drawn for indeterminate checkboxes (in pixels). |
 | roundingRadius | --gdg-rounding-radius | number | The radius of rounded corners used by various grid elements (in pixels). |
 
 If an option is missing from any theme it will be filled in with the default theme.